Restructure hero and nav
