+++ e-commerce/src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import BonusTop from './components/BonusTop';
-import NavBar from './components/NavBar';
+import BonusTop from './components/BonusTop/BonusTop';
+import NavBar from './components/NavBar/NavBar';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import Account from './pages/Account';
@@ -11,19 +11,16 @@ import Help from './pages/Help';
 
 function App() {
   return (
-
-    <p>Hello</p>
-
-      //   <BrowserRouter>
-      //     <BonusTop />
-      //     <NavBar />
-      //     <Routes>
-      //       <Route path={'/'} element={<Home />} />
-      //       <Route path={'/account'} element={<Account />} />
-      //       <Route path={'/cart'} element={<Checkout />} />
-      //       <Route path={'/help'} element={<Help />} />
-      //     </Routes>
-      // </BrowserRouter>
+        <BrowserRouter>
+          <BonusTop />
+          <NavBar />
+          <Routes>
+            <Route path={'/'} element={<Home />} />
+            <Route path={'/account'} element={<Account />} />
+            <Route path={'/cart'} element={<Checkout />} />
+            <Route path={'/help'} element={<Help />} />
+          </Routes>
+      </BrowserRouter>
   );
 }
