--- e-commerce/src/components/AutoSlide/sliderImage.jsx
@@ -1,41 +1,0 @@
-import First from "../../images/assets/First.jpg"
-import Second from "../../images/assets/Second.jpg";
-import Third from "../../images/assets/Third.jpg";
-
-export default [
-  {
-    title: "First Slide",
-    description: "This is the first slider Image of our carousel",
-    urls: First,
-  },
-  {
-    title: "Second Slide",
-    description: "This is the second slider Image of our carousel",
-    urls: Second,
-  },
-  {
-    title: "Third Slide",
-    description: "This is the Third slider Image of our carousel",
-    urls: Third,
-  },
-  // {
-  //   title: "Fourth Slide",
-  //   description: "This is the Fourth slider Image of our carousel",
-  //   urls: "https://mdbootstrap.com/img/Photos/Slides/img%20(133).jpg",
-  // },
-  // {
-  //   title: "Fifth Slide",
-  //   description: "This is the Fifth slider Image of our carousel",
-  //   urls: "https://mdbootstrap.com/img/Photos/Slides/img%20(147).jpg",
-  // },
-  // {
-  //   title: "Sixth Slide",
-  //   description: "This is the Sixth slider Image of our carousel",
-  //   urls: "https://mdbootstrap.com/img/Photos/Slides/img%20(105).jpg",
-  // },
-  // {
-  //   title: "Seventh Slide",
-  //   description: "This is the Seventh slider Image of our carousel",
-  //   urls: "https://mdbootstrap.com/img/Photos/Slides/img%20(102).jpg",
-  // },
-];
