+++ e-commerce/src/components/BonusTop/BonusTop.css
@@ -1,0 +1,28 @@
+*{
+    box-sizing: border-box;
+}
+body{
+    margin: 0;
+}
+@import url('https://fonts.googleapis.com/css2?family=Julius+Sans+One&family=Nunito+Sans:wght@300;400;700&family=Playfair+Display:wght@700&display=swap');
+
+
+.bonus-top{
+    text-align: center;
+    margin-bottom: 0;
+    font-family: 'Nunito Sans', sans-serif;
+    background-color: #FF7E07;
+    padding: 5px;
+}
+.bonus-top p{
+    margin: 0;
+}
+.bonus-top a{
+    color: white;
+}
+
+@media screen and (max-width: 768px) {
+    .bonus-top p{
+        font-size: 9pt;
+    }
+}
