+++ e-commerce/src/components/BonusTop/BonusTop.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './BonusTop.css';
 
 const BonusTop = () => {
   return (
