+++ e-commerce/src/components/Hero/Hero.css
@@ -1,0 +1,125 @@
+*{
+    box-sizing: border-box;
+}
+body{
+    margin: 0;
+}
+@import url('https://fonts.googleapis.com/css2?family=Julius+Sans+One&family=Nunito+Sans:wght@300;400;700&family=Playfair+Display:wght@700&display=swap');
+
+
+.hero{
+    background-image: linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4)), url("../../images/hero.png");
+
+    height: 500px;
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: cover;
+    position: relative;
+    width: 100%;
+    margin-bottom: 0;
+}
+.hero-content{
+    position: absolute;
+    top: 20px; 
+    left: 90px;
+    color: white;
+}
+.hero-content h1{
+    font-family: 'Playfair Display', serif;
+    font-weight: 700;
+    font-size: 60px;
+    line-height: 70px;
+    margin-bottom: 20px;
+    letter-spacing: 0.01rem;
+}
+.hero-content p{
+    font-family: 'Nunito Sans', sans-serif;
+    font-size: 20px;
+    line-height: 30px;
+    font-weight: 400;
+}
+.hero-buttons{
+    display: flex;
+    gap: 30px;
+}
+.hero-button{
+    font-family: 'Nunito Sans', sans-serif;
+    outline: 0;
+    border: 0;
+    border-radius: 60px;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    padding: 20px 0;
+    width: 200px;
+    text-align: center;
+    font-size: 12pt;
+    font-weight: 500;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    cursor: pointer;
+}
+.hero-button:hover{
+    box-shadow: 0 0 10px grey;
+    text-decoration: underline;
+}
+.hero-button1{
+    color: white;
+    background-color: #FF7E07;
+
+}
+.hero-button2{
+    background-color: white;
+    color: #FF7E07;
+}
+
+
+@media screen and (max-width: 768px) {
+    .hero{
+        background-image: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url("../../images/hero.png");
+    
+        height: 450px;
+        background-position: center;
+        background-repeat: no-repeat;
+        background-size: cover;
+        position: relative;
+        width: 100%;
+    }
+    .hero-content{
+        position: absolute;
+        top: 20px; 
+        left: 30px;
+        color: white;
+    }
+    .hero-content h1{
+        font-family: 'Playfair Display', serif;
+        font-weight: 700;
+        font-size: 30px;
+        line-height: 30px;
+        margin-bottom: 20px;
+        letter-spacing: 0.01rem;
+    }
+    .hero-content p{
+        font-family: 'Nunito Sans', sans-serif;
+        font-size: 16px;
+        line-height: 18px;
+        font-weight: 400;
+        width: 300px;
+        letter-spacing: 1px;
+    }
+    .hero-buttons{
+        flex-direction: column;        
+        gap: 10px;
+    }
+    .hero-button{
+        font-family: 'Nunito Sans', sans-serif;
+        outline: 0;
+        border: 0;
+        border-radius: 60px;
+        box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+        padding: 10px 0;
+        width: 200px;
+        text-align: center;
+        font-size: 11pt;
+        font-weight: 700;
+        box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+        cursor: pointer;
+    }  
+}
