+++ e-commerce/src/components/Hero/Hero.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './Hero.css';
 
 const Hero = () => {
   return (
